refactor: lotto에 대한 기능을 model에서 manager로 이동
